Updated to hierarchical configuration key
For `secrets.json`, `appSettings.json`, and **Azure Configuration**, the preferred convention is to use hierarchical values, which can be modeled using nested objects in JSON. This convention isn't supported in **OnTopic**, however, since the `:` hierarchy delimiter is a reserved character (for expressing the **OnTopic** hierarchy itself). As such, the delimeter is introduced for accessing variables from e.g. `secrets.json` or **Azure Configuration**, but then removed for accessing variables from **OnTopic**.