Introduced preview of *OnTopic Editor 4.0*
In parallel to the migration of **GoldSim** to **ASP.NET Core 3**, **Ignia** has _also_ been upgrading **OnTopic Editor** to **ASP.NET Core 3**—and setting it up so it can be distributed as a package. While the editor is not yet complete (there remain styling issues on the modal popups), this integration demonstrates how this will work, establishes all of the dependencies (e.g., in `Startup`, `GoldSimActivator`), and will make it easy to ugrade to the final release version, once it's available.