Sort technical papers by date
The technical papers have been updated to use a standardized date/time format (`yyyy-mm-dd`) and—in the new version of the **OnTopic Editor**, the new `DateTimeViewComponent`. This allows us to automatically sort the papers by `PublicationDate` instead of having to manually sort them in the editor (a tedious task). It also gives us more control to set the formatting across the board. To simplify all of this, I've created a method on the view model, `GetTechnicalPapers()`, which keeps the view simpler.