Introduced `EditorTemplate` for select box
There isn't any _implicit_ property type binding that makes sense for dropdown lists, which is why these were (previously) manually written out in each template. That said, we can _explicitly_ set the editor to use. This allows us to centralize the fromatting of dropdown (select) lists. Obviously, we still need to figure out how we're going to handle databinding of these, but that's a (mostly) separate issue.