Introduced CSS-based required label handling
In the legacy system, all required fields were explicitly prefixed with an asterisk. Preferably, this would be done with CSS. That is hard to accomplish, however, since **jQuery Unobtrusive** doesn't add an e.g. `required` class to `<label />` as it does with `<input />`. This can be mitigated by using the `ViewData.ModelMetadata.IsRequired` property—at least within an `EditorTemplate` bound to a single property (it's more involved to retrieve this value from a complex object). This allows us to reintroduce the original CSS selector for `label.required` which **bolds** required fields, while simultaneously adding support for prefixing these fields with an asterisk. We'll need to put more consideration into how to handle this for booleans, so that isn't (yet) implemented.