Reintroduced email domain validation
GoldSim has a business rule that emails must be submitted from institutional domains, not public domains. Strictly speaking, this is only _necessary_ for the _primary_ email address on the `Trial` and `Demo` forms. That said, unfortunately, **ASP.NET Core** doesn't recognize validation attributes that decorate overrides in derived classes, and so this rule must be placed on the base `CoreContact` class. That said, due to how the binding rules work, the corresponding action on `FormsController` must contain be bound explicitly to the field name (in this case, `BindingModel.Email`), so this won't successfully bind to any nested `CoreContact` objects (e.g. `TrialFormBindingModel.Trainer`). That conveniently fits with our current business requirements, but if those requirements change in the future we may need to reevaluate this approach.