Ensure `Forms` route is using the `FormsController`
This was getting picked up by the `MapContollers()` reference, but to ensure Ignia's `MapTopicRoute()` is being fully utilized, this should be necessary. Otherwise, it will use `TopicController` by default. (This works without this, but it'll likely introduce some ambiguity and confusion as we move toward integrating e.g. email templates and/or follow-up pages into the `/Forms` directory.)