Implemented `ITopicBindingModel`
I'd like to revisit this approach, but currently all overloads of `IReverseTopicBindingService` require that the models implement `ITopicBindingModel`—which in turn requires the implementation of `Key` and `ContentType`. This makes good sense for nested types in collections that are expected to be automatically mapped. It doesn't make sense when a topic is being supplied via the constructor. Regardless, it's a requirement for now, so this is being implemented in preparation for migrating to the `IReverseTopicMappingService`.