Rebased `PurchaseBindingModel` off of `ExtendedContact`
All of the other `FormBindingModel`s are fundamentally based off of `CoreContact` or a derived class (e.g. `ExtendedContact`, `ExtendedProfile`). The `PurchaseBindingModel` forms (e.g., `PurchaseFormBindingModel`) were not, however. The thinking behind this was that they had multiple contacts, so it made sense to treat each as its own property. But since the buyer contact is the primary contact, it makes more sense to maintain a consistent baseline data model. This will also simplify efforts should we ever wish to save this data to **OnTopic**, since the `FormData` content type is fundamentally based off of `CoreContact`.