Implemented the `IReverseTopicMappingService`
While not strictly necessary, migrating to the `IReverseTopicMappingService` reduces the code a bit and, more importantly, provides a lot more flexibility by allowing differences between how the object model is assembled and how that gets translated into attributes on the mapped topic (e.g., by using the `AttributePrefix` argument of the new `[MapToParent]` attribute). This required treating all generic type arguments representing the binding model to implement `ITopicBindingModel` and setting the `topicKey` and `contentType` on the `ITopicBindingModel` instance.