Introduce `MetadataLookupViewComponent`
The `MetadataLookupViewComponent` is an initial effort to handle lookup lists. It functions in that it's able to retrieve metadata based on model annotations using the `[Metadata()]` attribute. It has some serious limitations, however, in that it must be placed under an `Html.EditorFor()` or `<partial for />`, and it doesn't correctly wire up validation requirements. These will ideally need to be revisited at a future date.