Implemented `MetadataLookupViewComponent`
Implemented the newly introduced `MetadataLookupViewComponent` (#084dc99) on properties bound to lookup lists (#7b0c6e3) in the views. As part of this, got rid of `_Country.cshtml`, which was largely put in as a placeholder with the assumption that the country logic would require a more involved partial—that complexity has been shuffled off to the view component, however. That said, I kept the `Select.cshtml` editor template, however, as that allows us to keep the overall wrapper (e.g., markup, CSS, and label) in a consistent location as other fields. Further, it also mitigates a fundamental design flaw in the `MetadataLookupViewComponent` which otherwise prevents the fully qualified `HtmlFieldPrefix` from being properly calculated; by wrapping it in an `EditorFor()` it is able to inherit that value via the `ViewData`.