Removed `_PhoneNumber.cshtml` intermediary
It's not entirely clear why this was originally introduced—possibly in anticipation of more complex validation logic? Or maybe just to handle the unfortunate necessity to explicitly define that it should use the `string` editor template? Regardless, it doesn't provide much value, so we've replaced it with direct calls to `EditorFor()`.