Updated work with new `FormData` models
With the update of the forms from **ASP.NET WebForms** to **ASP.NET Core**, we've made a number of changes to the data model. This required updating the `ReportingController` (to look for the new content types) as well as the `ExcelReportingService` (to look for the new attribute names). While I was at it, I refactored the code to make it shorter and more maintainable. For instance, I introduced a new local `requestedModules()` helper function that greatly reduces the amount of code required to determine the configuration.