Refactored `Reporting` to `Licenses`
If we wanted to handle any other type of reporting, it's not clear that the `ReportingController` was going to meet that requirement, as it needs a number of actions that are license-specific. We may revisit that later, but for now it makes more sense to focus it in on it's primary use case. It also makes sense to base it off of `TopicController` so that any interface elements can be placed in the **GoldSim** interace; previously, they were in the **OnTopic** interface, which worked well from a design perspective, but is harder to depend upon now that it's included as a **NuGet Package**.