Established data model for license processor
The data model is a bit more involved than it strictly needs to be. By having a placeholder for each of the supported licensing types (e.g., `InstructorAcademicFormTopicViewModel`), though, we're able to automatically map child topics to a strongly typed topic view model, which we can expand in the future if we need additional properties. Further, because all of these inherit from `LicenseRequestTopicViewModel`, that can be used to implicitly filter `Requests` in the `LicenseAdministrationTopicViewModel`; it will evaluate _all_ children, regardless of depth, but only include those that implement `LicenseRequestTopicViewModel`.