Introduced basic license processor interface
Previously, while the `ReportingService` (now `LicenseExportService`) was handled via a controller (now `LicensesController`), the actual frontend interface was implemented via the legacy `LicenseRequests.aspx` **ASP.NET WebForms** page (ugh). This commit is an initial port of that over to **ASP.NET Core** using the **GoldSim** interface. (The previous implementation used the **OnTopic** interface, which looked nice, but doesn't make sense for a GoldSim-specific tool.) This required ripping out a lot of extranous markup required by **ASP.NET WebForms** and **Twitter Bootstrap** (which **OnTopic** uses). The CSS and JS were previously inline, and are now moved to separate `View` resources, which will be built as part of the new **Gulp** workflow, thus keeping the view a bit more focused on the markup. The class names and markup will still likely require a bit of care from @ktrunkey, but they are currently functional and looks more-or-less right.