Duplicated `favicon` in root
The main `_Layout.cshtml` specifies a location in `/Images/Icons` for the favorite icon. But paths accessing the site that aren't part of the `_Layout.cshtml` will still default to looking in the root for `favicon.ico`. This isn't strictly necessary, but is an easy fix for those edge cases.