Hack: Forced accept terms of use
There aren't any terms of use for requesting a demo, as the user isn't e.g. being opted in for the newsletter. But the model requires this. This is probably a bug in our model that we should reevaluate. Since every other form with a similar data model requires the `AcceptTermsOfUse`, however, we're hacking around this for now by hard-coding an accept value as a `hidden` input.