Introduce support for cancel buttons
Most forms won't need this. But we want to make it reusable. For now, the only form using it is the `Newsletter.cshtml`—which isn't exactly a "cancel" button, but rather an "unsubscribe" button. Contextually, though, they serve similar purposes. The "unsubscribe" button still needs to be wired up, but this establishes support for the user interface, at least.