Placed radio buttons outside of labels
Ideally, radio buttons would be placed inside of labels so that we don't need to explicitly define an `id` on the `<input>` and a `for` on the `<label>`. Unfortunately, this interferes with the plugin we're currently using to achieve fancy radio buttons. Consequently, it also interferes with the grid. Oh well. We'll probably want to revisit the plugin later, and likely go with a DIY/pure approach. Until then, this is a quick fix that resolves the issue.