Wrapped the `Boolean` editor template in a single column cell
Previously, every checkbox was wrapped in a single column (i.e., `medium-6`) width `cell` in the main templates. This was often done to group them together. By wrapping _every_ checkbox in a single column via the editor template, we not only centralize the code, but actually allow more flexible column wrapping for checkboxes. This is most notable with lists of checkboxes, as seen with e.g. modules and the advanced training options on the **User Conference** form, which will now flow naturally into a two-column format, without needing to be forced into it by grouping them into two individual columns.