Explicitly defined container dimensions
It seems that the default container dimensions in **Zurb Foundation 6** have changed, notably changing the margin to padding—thus eating into the previous width. To mitigate this, I explicitly set the width to `960px` (the width our site is designed for at full screen) and the padding to `0px`. I'm not certain this is the best approach, though, and this should be reviewed by @ktrunkey.