Fixed *Font Awesome* icon names
The **Font Awesome** icon names have changed between versions. Further, while **Font Awesome** _did_ introduce a new `fab` font ("b" is for "brands"), all of the file types we're selecting for are actually still in the normal `fa` (or `fas`) libraries (which is a bit unexpected, but convenient for us).