Allow forms to include an optional `RequestType`
When the `RequestType` is (optionally) passed to the `ProcessForm<T>()` method, it is appended to the email subject. This is useful if there are multiple request types on a form. For instance, the `Newsletter.cshtml` has an `Removal` request type. This value is defined on a "cancel" button (see #7b9c9d5) and bound to its action. That way, any code processing the email receipt can optionally respond to that request.