Worked around `MetadataLookup` issues
Ideally, we'd _just_ pass a `ModelExpression` to the `MetadataLookupViewComponent`, and that would be able to retrieve the `HtmlFieldPrefix`, `IsRequired` state, and `Value`, then reuse `asp-for` on the nested `<select />` list. This introduces a number of challenges, though. The `HtmlFieldPrefix` isn't set unless the view component is already nested inside of bound model (e.g. `@Html.EditorFor()`). And any model we bind to in the nested view will be used to determine e.g. `IsRequired`. That's unfortunate. Until we find a more elegant solution, we're mitigating this issue by explicitly setting the `HtmlFieldPrefix` on the view component, explicitly relaying the `Value` and `IsRequired` down to the view, and binding to the `HtmlFieldPrefix`, `Value`, and `IsRequired` instead of trying to bind to a nested model. This resolves the two remaining issues with the dropdown box—though, longterm, it'd be nice to revisit this issue and determine if there's a more elegant, general approach to solving this class of problems.