Introduced lookup lists for `_ProductSelection`
The `Quote` and `Purchase` forms require the customer to select which version of the product and which license type they want. These were being rendered as input boxes, requiring the customer to type these in. They _should_ be displayed as _dropdown_ lists, allowing the customer to _select_ a value. This update fixes that by binding them to a `Configuration:Metadata` list in the shared `PurchaseBindingModel`, and then using the new(ish) `MetadataLookupViewModel` in the shared `_ProductSelection.cshtml`. (Note that the latter actually calls the `Select.cshtml` editor template, which in turn calls the `MetadataLookupViewModel`; this allows the field to be wrapped in a standard template.)