Ensured all extensions were removed prior to readding them
It turns out there may, in fact, be `woff` support already, as this was throwing a configuration exception. We'll want to go through this list and remove any which are now shipped with the latest version of **IIS**. For now, however, this is consistent with the existing framework. This issue was inadvertantly introduced in #dece5d0, when, apparently, the `<remove />` for `woff` files was not reintroduced as part of the sorting. Whoops.