Introduced basic implementation of `/Invoices`
This includes an `InvoicesController` with two key actions: `Index()` and `Edit()`, alongside their respective view models (`InvoicePageTopicViewModel` and `EditInvoicePageTopicViewModel`) and views (`Index.chstml` and `Edit.cshtml`). Further, there's a child view model—`InvoiceTopicViewModel`—that's shared between `InvoicePageTopicViewModel` and `EditInvoicePageTopicViewModel` which represents the core topic that's being displayed and edited (it is wrapped the `PageTopicViewModel`s so that additional page-specific information, such as the page name, can be passed to the view). The `InvoicesController` is explicitly registered in the `Startup` since we don't otherwise have a default route for `/{controller}/{action}/{id}`; it is also added to the `GoldSimActivator` for dependency composition.