Introduced support for renaming invoices
While previously, users could rename invoices, this would not actually rename the `Key`, which maps to the `InvoiceNumber` by convention. Since the `Key` is how invoices are looked up, this would effectively orphan the invoice. Now, when an invoice is edited, the `Key` is updated if a) the `Key` isn't null (implying an existing record), and b) it doesn't match the `InvoiceNumber`.