Reestablished `InvoiceTopicViewModel` as a stand-alone class
Instead of inheriting `Key` and `Id` from `TopicViewModel`, these are reproduced here. This is necessary so that these can be a) marked as nullable, and b) not set to `[Required]`. Otherwise, new invoices can't be created since the `Id` and `Key` won't be set for new records. (Fundamentally, the `TopicViewModel` class isn't designed to be repurposed as a binding model. We may want to revisit that at some point, but for now it makes sense to continue allowing those fields to be required until we've considered all ramifications of modifying them.)