Centralized and generalized administration scripts, styles
Renamed `LicenseAdministration.(js|css)` to simply `Administration.(js|css)` and generalized the language and identifiers to work across multiple administration pages. This makes these scripts better suited to serve both `Licenses/LicenseAdministration.cshtml` as well as `Invoices/Index.cshtml`. As part of this, needed to change the script locations as well as some class names in the views themselves, and decentralize the initialization of `DataTable()` since it contains table-specific properties (such as `lengthMenu` and `order`) which are harder to generalize.