Introduced row hover/click functionality
This script allows an entire row (optionally excluding any `js-no-click` cells) to be clickable when that row is hovered, with the click action redirecting to a `data-href` attribute. This is much more elegant than having a link within the `<td>`, and more consistent with the **GoldSim** brand (not that it matters as much for these internal pages, but details!).