Moved `Licenses`, `Invoices` to new `Administration` area
To help organize the code, while also providing a more consistent administration experience, administration tools are being grouped under a new `Administration` area. This has a number of benefits. It helps keep the administrative models, controllers, and views out of the main website area. It also allows the administrative tools to share a common navigation, making it easier to jump between them. As part of this, the namespace for models changes (e.g., from `GoldSim.Web.Models.Invoices` to `GoldSim.Web.Administration.Models.Invoices`) as did the underlying topics they reference (e.g., from `Root:Invoices` to `Root:Administration:Invoices`). Further, for consistency, license requests are now placed directly in `Root:Administration:Licenses` instead of in a `Requests` container; this is similar to how `Invoices` are handled. This required updating to build 22 of **OnTopic 3.8.0**, which corrects some issues related to area routing and views.