Fixed route
With the introduction of areas (see #53bf056), a new route parameter was added—`id`—which allows us to generalize the pattern of details pages, while also baking them into the path. That said, to generalize this, it required renaming it from `invoiceNumber` to `id`. This reference was missed as part of that process.