Replaced legacy validation with `ModelState` validation
Now that we're using the `PaymentFormBindingModel` (see #3ae5764) we don't need to assemble a dictionary of `ErrorMessages` as part of the view model; we can just use the OOTB `ModelState`. Technically, we could have done this before, but it's even easier now that we can just call `Model.IsValid` to check the validation attributes. That said, we still need to keep the validation provided by **Braintree** as it does its own specialized validation of the credit card information, which we don't want to replicate.