Initial implementation of `VerifyInvoice`
The `VerifyInvoice()` action provides server-side validation of the `InvoiceNumber` and `InvoiceAmount`, ensuring they match the values provided by the `InvoiceTopicViewModel`. That said, we'll need to put some additional effort into this to better distinguish the error states; as is, it will return the same error twice if the `InvoiceNumber` is invalid since the `VerifyInvoice()` action is called for both `InvoiceNumber` and `InvoiceAmount`. For now, however, this establishes the concept.