Separated `VerifyInvoice()` into `VerifyInvoiceNumber()` and `VerifyInvoiceAmount()`
Previously, since there was only one `VerifyInvoice()` method, there was the possibility of the customer receiving duplicate error messages in the scenario where the `InvoiceNumber` was incorrect. This is because both `InvoiceNumber` and `InvoiceAmount` validated using the same function, and would thus both trip on the invalid `InvoiceNumber`. To workaround this, I separated it into two distinct validators—`VerifyInvoiceNumber()` and `VerifyInvoiceAmount()`. This creates a small amount of duplicate code, unfortunately, but provides a better user experience. (The code duplication is really small, though I _may_ throw it into a helper function later to avoid that; TBD.)