Renamed view models for consistency
Working to establish a more consistent naming convention for view models. While both of these view models do derive from and derive topics, they don't map to actual content types, and aren't intended to be mapped automatically using the `ITopicMappingService` (though they can and are mapped manually). To differentiate them, the `TopicViewModel` is renamed to `ViewModel`. Also, removed `Page` since that's usually implicit (most of the view models map to pages). Finally, explicitly labeled the index view model as `InvoiceList` to better differentiate it, and to indicate that it represents a collection of invoices.