Explicitly set `Administration` area to `RequireAuthorization()`
Currently, both administrative controllers in the `Administration` area are annotated with `[Authorize]`. We shouldn't depend on that for the `Administration` tools, however. To be safe, the entire `MapAreaControllerRoute()` is set to `RequireAuthorization()` to ensure that any controller we add to it will require authentication against either **Ignia** or **GoldSim**'s **Azure AD**.