Removed `OrderBy` on email receipt
Previously, when an email receipt was sent, the fields were ordered alphabetically. This had the benefit of ensuring a consistent order. But that order didn't map nicely to the form interface, nor to logical groupings of fields. To mitigate this, we're reverting this back to the original model, where the fields are delivered in the same order that they're received.