Formatted administrative buttons
Normally, there's a large amount of padding at the top of a `panel`, which works well for text. For an administrative page with buttons and tables, it feels imbalanced. To account for this, the margin on the buttons is reduced to center them between the top of the `panel` and the `table`. Ultimately, we may want to find a better approach that reduces the padding for these types of pages, instead of compensating for it on the elements (in this case, buttons).