Hardcoded route to `PayInvoice`
Our workaround for intercepting requests to other controllers and replacing them with more specific controllers in the `GoldSimActivator` doesn't seem to work in **ASP.NET Core**. The page is delivered just fine, but the `[Post]` action isn't routed to. To correct this, we need to establish an explicit route that takes precident over the `TopicController` route. This isn't an ideal approach, and we should revisit how these (rare) scenarios are handled. In the meanwhile, this mitigate the issue while maintaining the legacy URL.