Disabled XML transformations
We _want_ XML transformations. That said, currently, they are being applied _twice_—and causing duplicate entries. That's preventing the application from running. To mitigate this, we're disabling this to determine if that resolve the issue. E.g., this may not be necessary if these transformations are already occurring as part of the build process (which is probably what's going on).