Disabled removal of extraneous files
As feared, the `RemoveAdditionalFilesFlag` doesn't seem to honor the `<MsDeploySkipRules />` from the `csproj` file, even though we're explicitly using `webDeploy`. That's a concern. For now, we'll remove this feature so that continuous integration ignores these files. That said, this isn't optimal since it will result in since-deleted static files being left behind, so we'll ideally want to look further into this.