Migrated `<staticContent>` to `FileExtensionContentTypeProvider` middleware
The `StaticFileModule` isn't supported when using **ASP.NET Core** on **IIS* (and, thus, **Azure App Services**). Files that need to support downloading _must_ be configured explicitly in the `Startup`. (This change was previously made, but not committed into source control until we had an opportunity to investigate the compatibillity issue further. With the introduction of _continuous deployment_ for the staging site, however, it's necessary to commit.)