Removed unnecessary folders
There are a number of folders which were carved out as exeptions to the (**Visual Studio**) deployment process, but which it turns out **GoldSim** no longer needs, and which don't get (much) traffic. Those have now been deleted, and are no longer needed as exceptions in the deployment process. (There is one file in `/Emails` that is still needed, but it's part of the **git** repository, so will be covered still.)