Update to final release of *OnTopic 4.0.0**
The final release of **OnTopic Library 4.0.0** includes some (additional) breaking changes primarily related to a) the refactoring of `Ignia.Topics` to `OnTopic`, and b) the conversion of `*Url` properties on view models from `string` to `Uri`, thus requiring some additional care in working with them in the views. We'll need to do one final upgrade in a couple of days, once **OnTopic Editor 4.0.0** is available.