Bumped *Topics Editor* version
The latest patch includes a (_temporary_) workaround to a bug inadvertantly introduced previously, which causes conflicts between the **Razor Class Library** views and the **GoldSim** views. This will need a better long-term fix, but in the meanwhile this allows the **OnTopic Editor** to run on the current version of **GoldSim**, at least. We're expecting a long-term fix in the next couple of days.