Updated favorite
GoldSim is now using the main icon from the product as their company icon, instead of the legacy black circle with the gold lightening bolt which used to be part of their logotype lockup. This is now reflected as part of their website favicons. As part of this, I also increased the resolution of the Android icon (to 512x512) and renamed the `manifest.json` to the newer `site.webmanifest` format. This required registering the `.webmanifest` MIME type.