Moved `EntityFramework` to private assets
We're only using **Entity Framework** as a design-time dependency for dynamically generating form views based on models. This is a dependency introduced by the **Visual Studio** designer. Since it's not required by the application itself, moving it to private assets. This doesn't actually mean much since we're not redistributing the website as e.g. a **NuGet** package, but it at least helps signal the intent.