Alphabetized redirects
The links were roughly in alphabetical order—but not entirely. Updated them to be in actual alphabetical order—with the exception of the slideshow redirects, which I maintained in oder of `#Slide` number so they're easier to evaluate. Hopefully, this makes it (slightly) easier to find items in the list in the future.