Establish new `RetiredSubsites` rewrite map
The `RetiredSubsites` rewrite map is similar to the newly established `Shortcuts` map in that it exclusively handles redirecting single-folder paths that live in the root. Unlike the `Shortcuts` map, however, it will also match any children of those directories. So, for example, the `Forums` key will map e.g. `/Forums`, `/Forums/`, and `/Forums/Licensing/3045`. This is useful for retired sites where _any_ URL from the legacy site should redirect to a _single_ target location (the `value`).