Establish new `SubsiteMoves` rewrite map
The `SubsiteMoves` rewrite map is very similar to the recently established `RetiredSubsites` rewrite map (see #1c68ee6) except that it will maintain the child URL structure. As such, `/Help/Child/Folder` will redirect to `https://help.goldsim.com/Child/Folder`. As the name suggests, this is useful for subsites that have been moved to a new location but maintain the nested folder and file structure. By contrast, the `RetiredSubsites` rewrite map will collapse the folder structure down to a single URL, since it assumed that the legacy folder/file structure is no longer relevant or maintained.