Treated `/CustomerRequests` as a `RetiredSubsite`
Technically, `/CustomerRequests` was a topic namespace used for storing, as the name suggests, customer requests (such as trial or academic requests). It wasn't intended for direct customer access, and thus the redirect. This isn't strictly necessary to maintain since that location no longer exists, but to be safe we're rolling it up into a `RetiredSubsite` since that's effectively how it's being treated. That allows us to replace two keys with one, while also making the semantics clearer (i.e., this is moved as part of that section now being retired, not because we reorganized the content).