Migrated de facto shortcuts to `Shortcuts`
A number of the legacy `PageMoves` were _actually_ shortcuts—and either explicitly intended to be, or were _de facto_ operating as such. For instance, `Pay` is, and always was, a redirect to the payment processor. And while e.g., `Web` and `Home` did exist at one point, they work very similar to shortcuts in practice, by redirecting to a more specific URL. Regardless, while the more general rule for handling `PageMoves` works for shortcuts, the `Shortcuts` page map is more specific and descriptive of the behavior.