Removed redundant `Library` from `PageMoves`
There is already a preexisting rule under the `RetiredSubsites` page map which handles the `/Library` root and all children. Prior to it being established as a `RetiredSubsites`, it was handled as its own dedicated `rule` (see #84dd5a0). As such, the `PageMoves` version is—and has been—unnecessarily redundant and can be removed.