Moved `Home` from `Shortcuts` to `RetiredSubsites`
In #37ad0b8, I moved `Home` to `Shortcuts` since it was acting as a _de facto_ shortcut. That said, there were actually child pages under `/Home`—notably `Default.aspx` and `Home.asp`—which were implementing their own `PageMoves`. These can all be consolidated to treat `Home` as one of the `RetiredSubsites`—which, actually, is arguably a more accurate description for it anyway.