Organized rules and rewritemaps into logical groups
Added whitespace and comment headers in order to logically group both `<rewriteMap />`s and `<rule />`s, thus making it a bit easier to quickly scan the XML to get a sense of what goes where. This should also make it easier in the future to deterministically place items in their correct location.