Simplified application areas redirect
Previously, the three application areas—`BusinessSystems`, `EngineeredSystems`, and `EnvironmentalSystems`—were each explicitly spelled out as options. Since all of these end in `Systems` and there are no other legacy folders that end with `Systems` that we need to exclude from this rule, these can be simplified with a `([a-z]*)Systems` pattern.