Resolved invalid constructor issue with URI's
Previously, when a `LearnMoreUrl` (`Uri`) was null, a fallback was created. The only problem? The `Uri()` constructor doesn't accept just `#`, thus throwing an error. In fact, the behavior is exactly as we'd want _without_ this, and so it's entirely unnecessary. Ideally, this should never happen—as then it will link to the current page—but if it _does_ happen, we shouldn't throw an _exception_.