Introduced barebones `ErrorController`
By default, `MapTopicRoute()` uses the `TopicController`—which, in turn, always returns a status code of 200. As such, while `MapTopicRoute("Error")` was correctly mapping the `Root:Error` routes (e.g., `/Error/NotFound`) used by IIS's custom errors, it was not reporting the correct status code to the browser. This, in turn, interferes with the analytics from both **Azure Application Insights** as well as **Google Analytics**. Ideally, we'd like to revisit this functionality later as part of **ASP.NET Core**'s `UseStatusCodePagesWithReExecute()` and/or by reintroducing an `ErrorController` into **OnTopic**. As a temporary solution, however, this ensures that these pages are returning the correct status code.