Integrate new `Dispatch()` error handler
Implements the new `Dispatch()` error handler as part of the **ASP.NET Core** pipeline. This approach, while better integrated with **ASP.NET Core**, remains less than ideal, however, as the `Dispatch()` method does a `Redirect()` and, as a result, changes the URL associated with the 404. This doesn't provide clean analytics. As such, this should be considered experimental.