Fixed invalid URLs in the breadcrumbs
The breadcrumbs were using an antiquated and inaccurate way of extracting the web path from the current topic (view model), as it had been created prior to the `TopicViewModel` including the `WebPath` property. Notably, the previous approach failed to remove `Root` from the beginning, resulting in broken links. This update takes advantage of the new `WebPath` property, is easier to read, and produces the correct output.