Rename service connection name
Recently, GoldSim migrated many of their resources from a pay-as-you-go Azure subscription to the subscription subsidized as part of their MPS membership. As this included their App Service Plan, this necessitated recreating their service connection. As part of this, I've also given it a shorter, more generic name, in case we need to update it in the future.