Explicitly configuring use of .NET Core 3.1.x
This _shouldn't_ be necessary, but we're having problems with .NET Core 3.1 on the live server, resulting in an unexpected "HTTP Error 500.31 - ANCM Failed to Find Native Dependencies" error. Hopefully, by explicitly setting the framework version, we can get around this. That said, we may (additionally?) need to adjust how we're handling the deployment itself.