Explicitly established dependency for `SqlClient`
The `Microsoft.Data.SqlClient` _should_ be inherited as a dependency from `OnTopic.Data.Sql`—and, indeed, that works fine when restoring the project via **Visual Studio** or publishing from it. But that dependency isn't getting picked up correctly when deploying via **Azure DevOps Pipelines**, even though it worked fine with the previous **.NET Core 3.0** version. We'll need to figure that out later. For now, this should provide a temporary solution.