Enable removal of additional files when publishing
Ideally, we want to remove any orphaned or legacy files that are on the server from previous publishing processes. GoldSim, however, contains a number of folders which are managed independently of the web deployment, and we don't want those to be deleted. As such, in addition to configuring `RemoveAdditionalFilesFlag`, we're _also_ configuring a number of `-skip:obectName=dirPath,absolutePath=` arguments to bypass those folders. These are _also_ specified in the `GoldSim.Web.csproj` as `MsDeploySkipRules` but, unfortunately, while **Visual Studio** honors those settings, they aren't recognized by the `AzureRmWebAppDeployment` task and, thus, need to be duplicated here.